requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.34.0
colorama==0.4.6
comm==0.2.1
cryptography==41.0.4
cssselect2==0.7.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
django-ckeditor==6.7.0
django-environ==0.11.2
django-extensions==3.2.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-multiforloop==0.3.1
django-paypal==2.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
html5lib==1.1
idna==3.4
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
lxml==4.9.3
matplotlib-inline==0.1.6
mysqlclient==2.2.0
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdf2image==1.16.3
Pillow==10.0.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pypdf==3.16.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
#pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==1.26.16
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.11
gunicorn==21.2.0
